Updated tags.delete with bulk delete feature

+++ tenable/io/tags.py
@@ -145,24 +145,34 @@ class TagsAPI(TIOEndpoint):
             payload['description'] = self._check('description', description, str)
         return self._api.post('tags/categories', json=payload).json()
 
-    def delete(self, tag_value_uuid):
+    def delete(self, *tag_value_uuids):
         '''
-        Deletes a tag category/value pair.
+        Deletes tag value(s).
 
         :devportal:`tag: delete tag value <tags-delete-tag-value>`
 
         Args:
-            tag_value_uuid (str):
-                The unique identifier for the c/v pair to be deleted.
+            *tag_value_uuid (str):
+                The unique identifier for the tag value to be deleted.
 
         Returns:
             :obj:`None`
 
         Examples:
+            Deleting a single tag value:
+
             >>> tio.tags.delete('00000000-0000-0000-0000-000000000000')
+
+            Deleting multiple tag values:
+
+            >>> tio.tags.delete('00000000-0000-0000-0000-000000000000', '10000000-0000-0000-0000-000000000001')
         '''
-        self._api.delete('tags/values/{}'.format(
-            self._check('tag_value_uuid', tag_value_uuid, 'uuid')))
+        if len(tag_value_uuids) <= 1:
+            self._api.delete('tags/values/{}'.format(
+                self._check('tag_value_uuid', tag_value_uuids[0], 'uuid')))
+        else:
+            self._api.post('tags/values/delete-requests',
+                json={'values': [self._check('tag_value_uuid', i, 'uuid') for i in tag_value_uuids]})
 
     def delete_category(self, tag_category_uuid):
         '''
